Working with tailwind
Ok so, for now, the only way I could get this to work is to have tailwind run
seperately, and have it handle outputting our css file. Luckily, it seems it can
handle scss files, so maybe we can just do that? Even luckier, it seems to be
automatically pruning and tree-shaking, so the output CSS is not monumental. How
it knows, I have no idea.